refactor: [FE] 파일 업로드 다운로드 별도 함수로 분리
(#71)
